[FEAT] group commands in help message
The help message contained all commands in alphabetic order. With the
great number of commands this becomes quite unreadable. This patch shows
the commands in groups. The support for this feature in App::Rad is
currently not on CPAN and has to be patched in. Thus, we check whether
App::Rad already supports it with UNIVERSAL->can.