[RFK] move creating new hosts into App::Rad based CLI
We want to reduce the number of dependencies. On of the dependency we
want to get rid of is App::Cmd by moving all CLI commands to
App::Rad. This is done step by step for time reasons. This patch moves
creating new hosts from "tapper-testrun newhost" to "tapper host-new".