[FEAT] update testruns in CLI
updating a testrun was not supported in the old CLI with
tapper-testrun. For status and auto-rerun such an update makes sense at
least for testruns that are not scheduled yet. Therefore, this patch
implements an update command for testruns that allowes just these to
changes. Other changes may become useful to some time. I did not get any
requests for other options to change though and therefore, I'll defer
adding more change options.

Note: The module is called Tapper::CLI::Test even though it works on
testruns. The correct name Tapper::CLI::Testrun is still occupied by the
old CLI "tapper-testrun". Will change the module name as soon as the old
commands are fully transfered to App::Rad.