Testplan.pm - Implement cli flag --id
Although it being documented in the help, the cli flag --id did up
until now not did not have any influence on the data gathered when
listing testplans. This commit implements the --id cli flag by using
the specified IDs as provided like they were just gathered from the
database into @ids. This provides the rest of the testplanlist
subroutine with the necessary filter requested by the user.